Finished problem D. Finally! I was overflowing int, I needed to be using long long

+++ ACM_ICPC_Fourth_Contest/problemD.cpp
@@ -9,15 +9,15 @@ using namespace std;
 
 int main()
 {
-    int A, B;
+    long long A, B;
     char Sin[5000];
     bool S[5000];
 
-    while(scanf("%d %d",&A, &B)!=EOF)
+    while(scanf("%lld %lld",&A, &B)!=EOF)
     {
         scanf("%s",&Sin);
 
-        int n = 0, p = 0;
+        long long n = 0, p = 0;
 
         for(int i = 0; Sin[i]!='\0'; i++)
         {
@@ -31,17 +31,17 @@ int main()
             n++;
         }
 
-        int minCost = 0;
-        int numToMove = 0;
+        long long minCost = 0;
+        long long numToMove = 0;
         for(int i = p; i < n; i++)
         {
             if(S[i]) numToMove++;
         }
         minCost = numToMove*A;
-        int swapCost = A-B;
-        int swappingCost = 0;
+        long long swapCost = A-B;
+        long long swappingCost = 0;
 
-        int curr = p;
+        long long curr = p;
         for(int i = 1; i <= numToMove; i++)
         {
             while(!S[curr]) curr++;
@@ -58,7 +58,7 @@ int main()
 
             minCost = min(minCost,swappingCost+(numToMove-i)*A);
         }
-        printf("%d\n",minCost);
+        printf("%lld\n",minCost);
     }
 
 }
